refactor(core): split god-file modules
